add experimental oracle manager
it is important that we make the right thing easy and the wrong thing difficult,
especially when it comes to oracles.

In order to make sure all synthetix system compenents (including 3rd party) are always able to add common-sense
oracle checks to their systems, we provide `OracleManager`, a tool for loading oracle prices
and subscribing to potential issues with them.

Note that an "Oracle" is defined as a system which retrieves all or part of its reported data
from off-chain, so if we are to make custom oracles (like the Uniswap TWAP oracle on v2x),
a new "wrapper" oracle should be created which also exports the AggregatorV3 interface, and modifies
the oracle result in the intended way.

All synthetix ecosystem components needing an oracle value should use the provided oracle manager
functions. Within these functions, the oracle will be queried and its value run against a variety
of checks. Once the query is completed, the result will be returned along with an array of "incidents".
An incident represents a particular event of note which the oracle consumer might want to act upon.
For example, if an aggregator tracking a token price suddenly goes to 0, a "JumpDouble" incident will be
generated, which occurs whenever an oracle's price changes by more than a factor of 2. This incident is stored,
so that even after the oracle event has occured, other processes also using the oracle manager can see
that it has occured and when it occured and suspend or take other steps as necessary.

This system is *permissionless* and works with any aggregator v3 compliant interface. Simply
pass in the address of your oracle, and the system will automatically record data and report
any known issues.

Having a unified oracle management/checking sysetm has a number of considerable advantages:
* its possible to get a much more complete dataset of oracle behavior and reliability
(think of insurance companies sharing accident data, etc.), meaning that more accurate
oracle data can be delivered
* no need to create a new system of oracle verification every time. Simply subscribe to the
oracle incidents your code cares about, and any relevant data about the oracle will be
delivered to you
* reduces the risk of a 3rd party market with good intentions failing due to the author
not knowing the complexity of oracles and how to be protected from them.
* allows for the synthetix team to worry less about tweaking oracle parameters and making
them safer, and more on consuming the value and taking the appropriate precautions on
a case by case basis